:recycle: refactor: Padronizando nome dos pacotes.
