--- src/main/java/com/job/jmc/api/controller/DefaultController.java
@@ -1,16 +1,0 @@
-package com.job.jmc.api.controller;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping(value = "/")
-public class DefaultController {
-  @GetMapping
-  public ResponseEntity<String> defaultRoute() {
-    return ResponseEntity.status(HttpStatus.OK).body("Bem-vindo à jmc!");
-  }
-}
