+++ src/main/java/com/job/jmc/api/entities/BaseDbEntity.java
@@ -1,0 +1,20 @@
+package com.job.jmc.api.entities;
+
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.MappedSuperclass;
+
+@MappedSuperclass
+public abstract class BaseDbEntity<T, DTO> {
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private T id;
+  public T getId() {
+    return id;
+  }
+  public void setId(T id) {
+    this.id = id;
+  }
+  public abstract DTO toDto();
+}
