--- src/main/java/com/job/jmc/api/entity/ServiceTask.java
@@ -1,38 +1,0 @@
-package com.job.jmc.api.entity;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.job.jmc.api.controller.dto.ServiceTaskDto;
-import jakarta.persistence.Entity;
-import jakarta.persistence.ManyToMany;
-import java.util.List;
-
-@Entity
-public class ServiceTask extends  BaseDbEntity<Long, ServiceTaskDto> {
-  private String name;
-  private String description;
-
-  @ManyToMany(mappedBy = "serviceTasks")
-  @JsonIgnore
-  private List<Maintenance> maintenances;
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public String getDescription() {
-    return description;
-  }
-
-  public void setDescription(String description) {
-    this.description = description;
-  }
-
-  @Override
-  public ServiceTaskDto toDto() {
-    return new ServiceTaskDto(this.getId(), this.getName());
-  }
-}
