--- src/main/java/com/job/jmc/api/exception/ControllerAdviceExceptionHandler.java
@@ -1,20 +1,0 @@
-package com.job.jmc.api.exception;
-
-import com.job.jmc.api.entity.Vehicle;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class ControllerAdviceExceptionHandler {
-  @ExceptionHandler({
-      EntityNotFoundException.class
-  })
-
-  public ResponseEntity<String> handleEntityNotFoundException(RuntimeException ex) {
-    return ResponseEntity
-        .status(HttpStatus.NOT_FOUND)
-        .body(ex.getMessage());
-  }
-}
