+++ src/main/java/com/job/jmc/api/exceptions/ControllerAdviceExceptionHandler.java
@@ -1,0 +1,19 @@
+package com.job.jmc.api.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class ControllerAdviceExceptionHandler {
+  @ExceptionHandler({
+      EntityNotFoundException.class
+  })
+
+  public ResponseEntity<String> handleEntityNotFoundException(RuntimeException ex) {
+    return ResponseEntity
+        .status(HttpStatus.NOT_FOUND)
+        .body(ex.getMessage());
+  }
+}
