+++ src/main/java/com/job/jmc/api/repositories/BaseRepository.java
@@ -1,0 +1,8 @@
+package com.job.jmc.api.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.NoRepositoryBean;
+
+@NoRepositoryBean
+public interface BaseRepository<T, ID> extends JpaRepository<T, ID> {
+}
