+++ src/main/java/com/job/jmc/api/repositories/MaintenanceRepository.java
@@ -1,0 +1,8 @@
+package com.job.jmc.api.repositories;
+
+import com.job.jmc.api.entities.Maintenance;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MaintenanceRepository extends BaseRepository<Maintenance, Long> {
+}
