+++ src/main/java/com/job/jmc/api/repositories/VehicleRepository.java
@@ -1,0 +1,8 @@
+package com.job.jmc.api.repositories;
+
+import com.job.jmc.api.entities.Vehicle;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface VehicleRepository extends BaseRepository<Vehicle, Long> {
+}
