--- src/main/java/com/job/jmc/api/repository/BaseRepository.java
@@ -1,9 +1,0 @@
-package com.job.jmc.api.repository;
-
-import java.io.Serializable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.NoRepositoryBean;
-
-@NoRepositoryBean
-public interface BaseRepository<T, ID> extends JpaRepository<T, ID> {
-}
