--- src/main/java/com/job/jmc/api/repository/MaintenanceRepository.java
@@ -1,8 +1,0 @@
-package com.job.jmc.api.repository;
-
-import com.job.jmc.api.entity.Maintenance;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MaintenanceRepository extends BaseRepository<Maintenance, Long> {
-}
