--- src/main/java/com/job/jmc/api/repository/ServiceTaskRepository.java
@@ -1,8 +1,0 @@
-package com.job.jmc.api.repository;
-
-import com.job.jmc.api.entity.ServiceTask;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ServiceTaskRepository extends BaseRepository<ServiceTask, Long> {
-}
