--- src/main/java/com/job/jmc/api/repository/VehicleRepository.java
@@ -1,8 +1,0 @@
-package com.job.jmc.api.repository;
-
-import com.job.jmc.api.entity.Vehicle;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface VehicleRepository extends BaseRepository<Vehicle, Long> {
-}
