--- src/main/java/com/job/jmc/api/repository/WorkshopRepository.java
@@ -1,8 +1,0 @@
-package com.job.jmc.api.repository;
-
-import com.job.jmc.api.entity.Workshop;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface WorkshopRepository extends BaseRepository<Workshop, Long> {
-}
