--- src/main/java/com/job/jmc/api/service/VehicleService.java
@@ -1,10 +1,0 @@
-package com.job.jmc.api.service;
-
-import com.job.jmc.api.controller.dto.VehicleDto;
-import com.job.jmc.api.entity.Vehicle;
-import java.util.List;
-import org.springframework.stereotype.Service;
-
-@Service
-public class VehicleService extends BaseService<Vehicle, Long, VehicleDto> {
-}
