--- src/main/java/com/job/jmc/api/service/WorkshopService.java
@@ -1,9 +1,0 @@
-package com.job.jmc.api.service;
-
-import com.job.jmc.api.controller.dto.WorkshopDto;
-import com.job.jmc.api.entity.Workshop;
-import org.springframework.stereotype.Service;
-
-@Service
-public class WorkshopService extends BaseService<Workshop, Long, WorkshopDto> {
-}
